+++ cc/lookup_server/parameter_client/src/parameter_client.cc
@@ -24,7 +24,8 @@
 #include "cc/core/common/uuid/src/uuid.h"
 #include "cc/core/interface/async_context.h"
 #include "cc/core/interface/errors.h"
-#include "cc/public/core/interface/execution_result.h"
+#include "cc/public/core/interface/execution_result_macros.h"
+#include "cc/public/core/interface/execution_result_or_macros.h"
 #include "cc/public/cpio/proto/parameter_service/v1/parameter_service.pb.h"
 
 #include "cc/lookup_server/converters/src/string_converter.h"