+++ cc/lookup_server/scheme_validator/src/scheme_validator.cc
@@ -18,7 +18,8 @@
 #include "absl/strings/string_view.h"
 #include "cc/core/common/global_logger/src/global_logger.h"
 #include "cc/core/common/uuid/src/uuid.h"
-#include "cc/public/core/interface/execution_result.h"
+#include "cc/public/core/interface/execution_result_macros.h"
+#include "cc/public/core/interface/execution_result_or_macros.h"
 
 #include "cc/lookup_server/scheme_validator/src/error_codes.h"
 #include "protos/lookup_server/backend/sharding_scheme.pb.h"