+++ cc/lookup_server/service/src/hashed_lookup_task.cc
@@ -23,7 +23,6 @@
 #include "absl/time/clock.h"
 #include "cc/core/common/uuid/src/uuid.h"
 #include "cc/core/interface/async_context.h"
-#include "cc/public/core/interface/execution_result.h"
 #include "cc/public/core/interface/execution_result_macros.h"
 #include "cc/public/core/interface/execution_result_or_macros.h"